% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{calibrateImage}
\alias{calibrateImage}
\title{Calibrate image scale using a scale bar}
\usage{
calibrateImage(file, units = "mm")
}
\arguments{
\item{file}{image filename}

\item{units}{real-world units}
}
\description{
Calibrate image scale using a scale bar
}
