% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageCalibrate.R
\name{calibrateScale}
\alias{calibrateScale}
\title{Calibrate image scale using a scale bar}
\usage{
calibrateScale(x, units = "mm")
}
\arguments{
\item{x}{A geoImage object.}

\item{units}{Real-world units of the scale bar.}
}
\value{
A calibrated geoImage object.
}
\description{
Select two points corresponding to the ends of a scale bar.
}
