% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoImage.R
\name{.new_geoImage}
\alias{.new_geoImage}
\title{Create a geoImage object}
\usage{
.new_geoImage(image, file = NULL)
}
\arguments{
\item{image}{image array}

\item{file}{filename}
}
\value{
object of class geoImage
}
\description{
Create a geoImage object
}
\keyword{internal}
