% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ht13}
\alias{ht13}
\title{Feldspar porphyroclasts in greenstone}
\format{
a magick-image object.
}
\usage{
ht13
}
\description{
A photograph of a mylonitic greenstone with feldspar porphyroclasts
}
\references{
Stephan et al. (2025)
}
\keyword{datasets}
