% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageEllipse.R
\name{imageEllipse}
\alias{imageEllipse}
\title{Digitize multiple ellipses on an image}
\usage{
imageEllipse(x, nmin = 5)
}
\arguments{
\item{x}{geoImage object}

\item{nmin}{minimum points per ellipse}
}
\value{
ellipse_collection object
}
\description{
Digitize multiple ellipses on an image
}
