% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageOpen.R
\name{imageOpen}
\alias{imageOpen}
\title{Open an image for geological measurements}
\usage{
imageOpen(file)
}
\arguments{
\item{file}{image filename}
}
\value{
geoImage object
}
\description{
Open an image for geological measurements
}
