% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTransform.R
\name{inverseTransformCoordinates}
\alias{inverseTransformCoordinates}
\title{Transform corrected coordinates back to original image}
\usage{
inverseTransformCoordinates(x, coords)
}
\description{
Transform corrected coordinates back to original image
}
