% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{locatorEllipseImage}
\alias{locatorEllipseImage}
\title{Fit ellipse interactively on an image}
\usage{
locatorEllipseImage(file, nmin = 5)
}
\arguments{
\item{file}{image filename}

\item{nmin}{minimum number of points}
}
\description{
Fit ellipse interactively on an image
}
