% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoImage.R, R/photo.R
\name{perspectiveCorrect}
\alias{perspectiveCorrect}
\title{Correct perspective distortion of a geoImage}
\usage{
perspectiveCorrect(file, width = NULL, height = NULL, output = NULL)

perspectiveCorrect(file, width = NULL, height = NULL, output = NULL)
}
\arguments{
\item{x}{geoImage object}
}
\value{
corrected geoImage
}
\description{
Click four corners clockwise.

Click four corners clockwise.
}
