% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{plotImage}
\alias{plotImage}
\title{Display an image with pixel coordinates}
\usage{
plotImage(file, asp = 1)
}
\arguments{
\item{file}{image filename}

\item{asp}{aspect ratio}
}
\value{
image array invisibly
}
\description{
Display an image with pixel coordinates
}
