% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageRotate.R, R/photo.R
\name{rotateImage}
\alias{rotateImage}
\title{Rotate image using a reference horizontal line}
\usage{
rotateImage(file, output = NULL)

rotateImage(file, output = NULL)
}
\arguments{
\item{x}{A geoImage object.}
}
\value{
Rotated geoImage object.
}
\description{
Select two points that should define a horizontal line.
}
