% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTransform.R
\name{transformCoordinates}
\alias{transformCoordinates}
\title{Transform coordinates of a geoImage}
\usage{
transformCoordinates(x, coords)
}
\arguments{
\item{x}{geoImage object}

\item{coords}{two-column matrix}
}
\description{
Transform coordinates of a geoImage
}
